requirements.txt
# ── Discord bot ───────────────────────────────────────────────
discord.py>=2.4.0
PyNaCl>=1.5.0

# ── Music ─────────────────────────────────────────────────────
yt-dlp==2024.4.9
spotipy==2.23.0
wavelink==3.3.0

# ── AI ────────────────────────────────────────────────────────
anthropic==0.25.8
openai==1.30.5

# ── Web / Dashboard API ───────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
gunicorn==22.0.0

# ── Database ──────────────────────────────────────────────────
pymongo==4.7.3

# ── Utils ─────────────────────────────────────────────────────
python-dotenv==1.0.1
aiohttp==3.9.5
Pillow==10.3.0

# ── TTS ───────────────────────────────────────────────────────
gTTS==2.5.1          # Text-to-Speech (Google TTS, free, no API key)

# ── Minecraft Bridge ──────────────────────────────────────────
mcrcon==0.7.0        # RCON client for Minecraft server communication
